requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorlog==6.7.0
cryptography
Deprecated==1.2.14
exceptiongroup==1.1.2
fake-useragent==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-translate==3.11.2
googleapis-common-protos==1.59.1
googletrans==4.0.0rc1
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna
importlib-resources==6.0.0
lxml==4.9.3
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
pdfminer.six==20221105
pdfplumber==0.10.1
phonenumbers==8.13.22
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.0
PyPDF2==3.0.1
pypdfium2==4.18.0
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
stem==1.8.2
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3
uvicorn==0.23.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.1
